translation_system/backend_v2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
pyyaml==6.0.1
python-multipart==0.0.6
pydantic==2.5.0